requirements.txt
Werkzeug==2.0.2
redis==4.1.0
Flask==2.1.1
psycopg2-binary==2.9.3
